+++ src/main/java/step3/views/InputView.java
@@ -1,4 +1,4 @@
-package step3.ui;
+package step3.views;
 
 import java.util.NoSuchElementException;
 import java.util.Scanner;
